% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/balancing.R
\name{balancing}
\alias{balancing}
\title{Balancing algorithm via Maximum Likelihood}
\usage{
balancing(param1, param2, sign, dist = rep("Normal", length(param1)),
  optimize = "solnp", lbounds = rep(0, length(param1)), ubounds = rep(Inf,
  length(param1)), forceInitialConstraint = TRUE, tol = 1e-05,
  constrTol = 1e-06, plot = FALSE, plotSigma = 3)
}
\arguments{
\item{param1}{A vector of the first parameter for each of the elements.  For
a normal distribution, this is the mean.}

\item{param2}{A vector of the second parameter for each of the elements. For
a normal distribution, this is the standard deviation.}

\item{sign}{A vector of the sign of each element.  These values should all
be +1 or -1, and they indicate if Delta_1, Delta_2, ... should be
pre-multiplied by a negative or not.  Usually, these will all be +1.}

\item{dist}{A vector of the name of the distribution for each distribution.
Currently, only "Normal" is implemented. Working in progress for different
distribution}

\item{optimize}{A string with the method of optimization of the Maximum
Likelihood, default solnp (Rsolnp dependency).  Must be one of "solnp",
"L-BFGS-B" (uses optim and is NOT recommended, as it hasn't been tested
thoroughly and doesn't enforce all constraints), or "constrOptim" (using
constrOptim from base package).}

\item{lbounds}{A Vector of the lower bounds for each element. These values
should all be numeric}

\item{ubounds}{A Vector of the upper bounds for each element. These values
should all be numeric}

\item{forceInitialConstraint}{Should the initial parameter vector be forced
to satisfy the constraints?  If TRUE, the initial value for one element
will be adjusted in an attempt to satisfy the constraints, but this may
not be possible.  If not, a warning is given and optimization proceeds
with the provided starting value.}

\item{tol}{The level of tolerance of the balancing (numeric). By default is
set up to 1e-5.  Currently only used if all elements are fixed, and in
this case it provides the numerical tolerance requried to check if the
balance is satisfied.}

\item{constrTol}{The tolerance to pass to the constrOptim algorithm.  This
algorithm does not force that the equation is balanced but rather forces
that the equation is within +/- constrTol % of being balanced.  For
example, if constrTol = 0.0001, then a tolerable error of .01% is allowed
in the balance.  This should be set very small as later any lack of
balance is assigned to food, and such assignment could create negative
values.}

\item{plot}{Logical.  Should a plot be generated which shows the result of
the balancing?}

\item{plotSigma}{A plotting parameter.  See ?plotBalancing.}
}
\value{
A vector of the final balanced values
}
\description{
This function balances the data using maximum likelihood.
}

