% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defaultStandardizationParametersBalancing.R
\name{defaultStandardizationParametersBalancing}
\alias{defaultStandardizationParametersBalancing}
\title{Default Standardization Parameters Balancing}
\usage{
defaultStandardizationParametersBalancing()
}
\value{
A list with the standardization parameters for the balancing.
}
\description{
Provides an object which contains the standardization parameters.  This
allows for easy passing into functions.
}

